+++ src/coro_rdma/tests/test_device.cpp
@@ -1,4 +1,4 @@
-#include <ylt/coro_rdma/ib_device.hpp>
+#include <ylt/coro_io/ibverbs/ib_device.hpp>
 
 #include "doctest.h"
